fix: minor init.sql changes
added username to user: split tags into tags and wine_tags where tags is the table of reusable tags and wine_tags is the association to a wine

+++ backend/internal/db/storage_location.go
@@ -5,7 +5,7 @@ import (
 	"database/sql"
 )
 
-func CreateStorageLocation(db *sql.DB, ctx context.Context, userID int, name string) (*StorageLocation, error) {
+func CreateStorageLocation(db *sql.DB, ctx context.Context, userID int64, name string) (*StorageLocation, error) {
 	location := &StorageLocation{}
 	err := db.QueryRowContext(ctx, `INSERT INTO storage_locations (user_id, name) VALUES ($1, $2) RETURNING id, user_id, name`,
 		userID, name).Scan(&location.ID, &location.UserID, &location.Name)
@@ -15,7 +15,7 @@ func CreateStorageLocation(db *sql.DB, ctx context.Context, userID int, name str
 	return location, nil
 }
 
-func DeleteStorageLocation(db *sql.DB, ctx context.Context, id int) error {
-	_, err := db.ExecContext(ctx, `DELETE FROM storage_locations WHERE id = $1`, id)
+func DeleteStorageLocation(db *sql.DB, ctx context.Context, id int64, userID int64) error {
+	_, err := db.ExecContext(ctx, `DELETE FROM storage_locations WHERE id = $1 AND user_id = $2`, id, userID)
 	return err
 }
